Added support for Basic Authentication and support latest Nexus3
Added ability to download artifacts from repositories protected with
Basic Authentication. In this case, the username and password can be
specified in `credentials.properties` and they are used against the API
and artifact URLs.

Also the app has been fixed to work with the current version of Nexus3:
3.20. There were changes in the REST API url needed to get it working
again.

Finally, updated Spring Boot and Guava to their latest versions and
removed the Spring Beta repositories, as we're now using the stable
version of Spring Boot 2.2.4.